# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPAY API",
        "description": "Documentation OpenAPI pour l'API OMPAY. Application de paiement mobile avec authentification par SMS, transferts d'argent et paiements QR.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Environnement local"
        }
    ],
    "paths": {
        "/api/auth/envoyer-lien": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Envoyer un lien d'authentification",
                "operationId": "17f81560c708ec96b34131e426c01f56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "771234567"
                                    },
                                    "redirect_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "http://localhost:3000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lien envoyé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lien de connexion envoyé si le numéro existe."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/echange": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Échanger un code d'authentification contre un token et données utilisateur",
                "operationId": "22835a2c8a60e3ad7704be14dfea6136",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "temp_token"
                                ],
                                "properties": {
                                    "temp_token": {
                                        "type": "string",
                                        "example": "abc123..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie avec données complètes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "phone": {
                                                            "type": "string"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "is_phone_verified": {
                                                            "type": "boolean"
                                                        },
                                                        "compte": {
                                                            "$ref": "#/components/schemas/Compte"
                                                        },
                                                        "qr_code": {
                                                            "$ref": "#/components/schemas/QrCode"
                                                        },
                                                        "recent_transactions": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/Transaction"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authentification réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Code invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Compte": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float"
                    },
                    "devise": {
                        "type": "string",
                        "example": "XOF"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "QrCode": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "code": {
                        "type": "string"
                    },
                    "meta": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "compte_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "transfert",
                            "paiement"
                        ],
                        "example": "transfert"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "type": "string",
                        "example": "completed"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}